Remove outdated Debian packaging code.
There are now official Debian and Ubuntu packages for open-iscsi that
were originally based on this code but have been updated with new
versions. The current packaging code on the other hand here hasn't been
updated in 10 years and is out of date (probably doesn't even build
anymore).

Remove it, this is better maintained downstream in the respective
distributions, especially when it comes to integration with the boot
process.