iscsiuio: Make builds reproducible if SOURCE_DATE_EPOCH is set
Build systems that capture the build date do not create reproducible
binaries. If the builds are reproducible on a bit level, the build time
stamp is actually irrelevant. A much better time stamp would be the
last change of the source package. For this reason, a standard was
devised that this date could be passed in externally to build systems:
the environment variable SOURCE_DATE_EPOCH could contain the UNIX
timestamp of the source package's last modification, and the build
system could use that instead of the current date and time.

On Debian systems (but there are also efforts to make e.g. Fedora
reproducible) the Debian tooling would pass SOURCE_DATE_EPOCH into the
build system.

Alter iscsiuio/Makefile.am and iscsiuio/configure.ac to use that
environment variable if set instead of the current date to create
src/unix/build_date.c.

Additionally, enforce UTC as time zone when SOURCE_DATE_EOPCH is set,
to ensure that the user's time zone has no bearing when trying to build
a binary that's supposed to be reproducible.

See also:

https://reproducible-builds.org/docs/timestamps/
https://reproducible-builds.org/docs/timezones/
https://reproducible-builds.org/specs/source-date-epoch/

Signed-off-by: Christian Seiler <christian@iwakd.de>