delete unused BSD stub code
This never supported any BSD drivers, they have their own tools, this is
just old unused stub code.  This is a Linux only project.

The cleanup of usr/Makefile also stops checking for kernel version.
NETLINK_ISCSI should come from a kernel supplied netlink.h, and should
always be 8.  The values of 11 or 12 I think were speculating on what
the assigned value might be when the iSCSI code was still out of tree,
it's been assigned as 8 since 2005.