libopeniscsiusr: setup ipv6 records based on iface name
The is_ipv6 flag (runtime iface rec representation only, not on disk)
was being set when creating ifaces from sysfs, but not when reading them
in from an iface record file. This caused ipv6 iface records to be
reported as malformed and ignored.

Fix that with a name check for an "ipv6" substring. Old code also looked
at other fields, this is enough to fix iface records created from sysfs
for be2iscsi and qla4xxx. Further fixes might be required for other IPv6
configurations.