iscsiuio: fail on nic_nl_open failing
The return value of nic_nl_open was ignored and while working by chance
being the latest call before the error label any later insertion of code
after this might break it.

Avoid that by checking for the retval and explicitly jumping to the
error label in case it is non zero.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>