Make reconnect to session on startup forever default.
The daemon has traditionally tried to recover stale
sessions, at startup, in serial and in a forked process.
A recent commit, b9afe4709900, added the ability to limit
these retries, and it made the default 32. Commit
467946a945cc added back in the ability to retry forever,
if the retry count was set to 0, but left the default
at 32. This commit changes the default to 0, so as
not to disrupt the expected reconnection behavior by
default.