Use sd_notify() to tell systemd when iscsid is ready.
In addition, to be able to tells systemd when we are
reloading sessions, the code around forking a copy and
restoring sessions gets its PID saved so that we know
when it is done and can update systemd status.

Note that this requires systemd unit file iscsid.service
to use Type=notify and NotifyAccess=main.