Fix i586 build issues with string length overflow.
Gcc7 warns of possible string print overflow, on i586,
when printing password length (via a macro), generating
errors like:

[   59s]   ^~~~~~~~~~~~~~~~~~~~
[   59s] In file included from /usr/include/stdio.h:862:0,
[   59s]                  from idbm.h:27,
[   59s]                  from context.h:22,
[   59s]                  from idbm.c:59:
[   59s] /usr/include/bits/stdio2.h:64:10: note:
'__builtin___snprintf_chk' output between 2 and 11 bytes into a
destination of size 8
[   59s]    return __builtin___snprintf_chk (__s, __n,
__USE_FORTIFY_LEVEL - 1,
[   59s]           ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
~~~~~~~~~~
[   59s]         __bos (__s), __fmt, __va_arg_pack ());
[   59s]         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
[   59s] cc1: all warnings being treated as errors
[   59s] make[1]: *** [<builtin>: idbm.o] Error 1
[   59s] make[1]: Leaving directory

The fix is to limit the size of the string printed, so that no
overflow is possible.

The print macros in usr/idbm.c were updated, as well, to match
the newer version in libopeniscsiusr/idbm.c, also to help the
i586 build.