Fix output for iscsiadm node/iface print level P1
The output from 'iscsiadm -m node -P1' and 'iscsiadm -m iface -P1'
has regressed with the introduction of libopeniscsiusr.

The pre-lib output for 'iscsiadm -m node -P1' before was:

>Target: <TARGET-IQN>
>        Portal: <IP-ADDRESS>:<PORT>,1
>                Iface Name: default

repeated for each Node in the node datbase, and the output
of 'iscsiadm -m iface -P1' before the change was:

>Iface: default
>        Target: <TARGET-IQN>
>                Portal: <IP-ADDRESS>:<PORT>,1

repeated once for each Node in the database that uses the
'default' interface (i.e. usually all nodes).

In both of these cases, the output contained details for each
and every node in the database. But this evidently wasn't the
intention if one looks at the code. The code appears to try
to only print the details for each 'Target' if and only if
those details are different than the previous line. In other
words, the code looks like it wants to suppress the 'Portal'
and 'Iface Name' lines if they are the same as the previously-
listed target. This explains why libopeniscsiusr implemented
the output of iscsiadm to suppress this info on all but the
first target, assuming the Portal and IfaceName were the
same.

The fix is to simplify print_nodes_tree() to print details
for every Node, since that is how the output worked
before the libopeniscsiusr update.