iscsiadm: fix host stats mode coredump
The command:

> iscsiadm -m host -C stats -H <mac_addr>

Copies in the user-supplied mac address without checking its
length, but the maximum length is 17 characters. This overflows
the next field in the structure, or if libc checking is
enabled, produces this error:

> *** buffer overflow detected ***: iscsiadm terminated

Fix this by checking the length of the supply MAC address
before copying it into the array. Some extra debugging
error messages were added in the process.