Add etc/systemd/iscsi-init.service to SYSTEMDFILES Makefile variable
iscsi-init.service is not included in the SYSTEMDFILES Makefile variable,
causing it to be skipped when the install_systemd rule is executed. This causes
the iscsi systemd service to fail on a fresh install, as it requires the
iscsi-init service.

This commit adds etc/systemd/iscsi-init.service to the list of files in
SYSTEMDFILES to correct this issue.