libopeniscsiusr: skip over removed sessions
When looping over all sessions with iscsi_sessions_get, it's possible to
race against sessions being destroyed and have the sysfs attribute files
be removed before they're read. Let's not treat this as an error, and
simply drop the session that failed to read from the list. I think it
makes sense to treat session that disapear while they're being read as
if they were already gone when the sessions directory was first scanned.

Apparently having iscsiadm exit with an error when trying to get a list
of sessions is a problem for OpenStack deployments.

Signed-off-by: Chris Leech <cleech@redhat.com>