iscsistart: Fix iscsistart timeout during login session
When iscsistart places a login request during boot, if it gets timeout
and response receive after it. It triggers "connection1:0: detected conn
error (1011)", since in userspace nobody is listening.

On OL6, it has been observed where iscsid is not part of initramfs.

During boot, iscsistart forks and child process used as placing a request whereas
parent wait in event loop.

In the child, login_seesion places login request, and request gets
timed-out then parent also exit. If it receive response for timed-out
request and during that time no parent is listening to it then it will
error as 1011.

This patch adds an additional option to increase tmo for login request
timeout. This patch has been tested some intesive workload by increase it as 6x
and helps resolve the issue.

Signed-off-by: Jitendra Khasdev <jitendra.khasdev@oracle.com>