iscsi sysfs: check state before onlining devs
In 5.6, the commit:

commit 0ab710458da113a71c461c4df27e7f1353d9f864
Author: Bharath Ravi <rbharath@google.com>
Date:   Sat Jan 25 01:19:25 2020 -0500

    scsi: iscsi: Perform connection failure entirely in kernel space

made it so the kernel can start the recovery process. This means that
after the start conn operation the kernel could set the device into the
block stated. We can then hit a race where iscsid has done start conn,
and is calling session_online_devs but the kernel has hit an issue and is
now setting the device's to blocked.

This adds a check for if the device is in the offline state before trying
to set the state to running.