Update the iscsi-gen-initiatorname script: harden and generalize
The iscsi-gen-initiatorname script seems only to be used by
the SUSE distrubution, so it has never been very "general". This
update makes the script more resiliant, with better error
messages if a problem occurs, as well as more generic, allowing
the caller to set the "base IQN" value, though the default
contains to be the same.

Changes:
- Added use of getopts to parse options
- Added a "-h" option for help
- Added an option to supply base IQN
- Added better checking of options and params
- Added comments to the script itself
- Now check to see if iname file can be written to
  (handling read-only mounts better)
- Fixed initiator name listed in file (it was wrong)
- Removed use of printf(1)
- Fixed iBFT initiator name setting in general
- Now set iname file mode to 0600 even when it comes from
  iBFT
- Cleaned up the script to use variables instead of repeating
  file pathnames or other values