iscsistart: fix login timeout handling
We should set the timeout to -1 for the login request becuase if we've
connected to our event loop then the login will either complete, fail
and hit the login retries limit in the initiator, or we will crash and
we will exit. Either way we are going to return evetually, and the
user can configure the retries so they can also limit the time they
will wait.

If we've sent a login request to our event loop we also don't want to
send another one because the initiator will see it on its session loop
and just return immediately.

Signed-off-by: Mike Christie <michael.christie@oracle.com>