actor: add name to struct actor and init it with function name
Previous actor log makes it hard to figure out what
is going on in the thread, such as:

"deleting a scheduled/waiting thread!"

We only know a thread is deleted, while no other info
like what this thread is about to do and which thread
it is. Logs like following seems better:

"deleting a scheduled/waiting thread 01111e48: session_conn_error !"

It tells thread with id 01111e48, which is going to executed
function "session_conn_error" is deleted.

The commit tries to give name to each thread, the name
is going to be printed with previous messages to make
the log easy to read.

Signed-off-by: Wenchao Hao <haowenchao@huawei.com>