Do not allow NOP-outs when running in initrd
When open-iscsi is used for the root disc, it's
run at boot time in the initrd/initramfs. As the
documentation for open-iscsi discusses, NOPs are
a bad idea for this case. But it's difficult for
users to ensure NOPs are off when runnig in this
environment.

Fix this bug ensuring that NOPs are off (set to
zero) when running in initrd by checking for the
file /etc/initrd-release.