Fix iscsi-init.service start time
The commit e6e53d6e60bf ("Fix iscsi-init so that it runs when
root writable") added "After=root.mount" to iscsi-init.service,
but that was an error, since that means waiting for "/root".

Remove that line and add the proper configuration so that
the iscsi-init service does not try to write to
/etc/iscsi/initiatorname.iscsi unless /etc/iscsi is
writable.