iscsid: simplify qtask handling
We only have qtasks for login/sync or logout. There is no need to pass
it around when we can just have 2 pointers for the 2 types. We can then
quickly check if a login or logout is in progress and in the future
kill a running login when we get a logout request.

Signed-off-by: Mike Christie <michael.christie@oracle.com>