Allow name/value args for firmware logins.
This adds 2 things to the firmware login mode: first, it adds
sane default values for NOP-OUTs (off), login retry max, and
type of scan. Lastly, it also gives the caller the ability
to set name/value pairs on the command line, which gives the
ability to override these new defaults if desired.

The output of "iscsiadm -h" was also cleaned up a bit, removing
duplicate spaces.