Use kernel initiatorname when setting local iname
Update the iscsi-gen-initatorname script to take its
initiator name from the kernel command line, if present.
Also, beef up and simplify the error checking. Error
messages will now go to stderr, and it is now an
error to overwrite the initiatorname file unless
the "force" option is set.

The man page was updated, as well.