Create an systemd iBFT rule generator
When dracut is booted with 'ip=ibft' it will rename
the interfaces to "ibft*". This will cause error messages during
booting, as later on systemd is trying to rename the ianterfaces
again, which doesn't work as the interface is busy.
So create a systemd generator for creating a dynamic
udev rule to stop udev from renaming that interface.