using ISCSI_UEVENT_DESTROY_SESSION_ASYNC to destoy session
The netlink assumed operations with kernel are fast, so the implement of
__kipc_call() is a sync operation. It would wait for kernel's response
once it sent the ISCSI_UEVENT_XXX.

While the ISCSI_UEVENT_DESTROY_SESSION might takes a long time to get a
response from kernel, which would block iscsid.

Following is a scenario:

When connection is not in ISCSI_CONN_STATE_LOGGED_IN, session_unbind()
would not be called. So when ISCSI_UEVENT_DESTROY_SESSION is sent to
kernel, the kernel might perform unbind session, which is slow in
session with a massive number of devices.

So here we using ISCSI_UEVENT_DESTROY_SESSION_ASYNC other than
ISCSI_UEVENT_DESTROY_SESSION to address it. Kernel would execute
session destroy async.

Signed-off-by: Wenchao Hao <haowenchao@huawei.com>