iscsid: fix logout pdu send failure handling
This fixes a bug where when we send logout pdu we are incorrectly
handling it as an error. This was added in

https://github.com/open-iscsi/open-iscsi/pull/317

I think there might have been a mistake for which version of that patch
got merged because there are some unhandled review comments in the merged
code.

The patch also fixes some review comments that were given in that review
but were not handled when it was merged:
1. No space before "Dropping session"
2. Don't split error messages.