iscsid: destroy session with no connection when restart
A session might have no leadconn in next scenario:

- iscsid died after destroy_conn() is called during logout
  now the session state could be LOGED_IN or FREE
- iscsid died after create_session() is called during login
  now the session state would be FREE

For the first scenario, kernel is waitting for userspace's
destroy_session request, iscsid should perform this destroy
operation when restart.
While for the second scenario, iscsi_sysfs_get_sessioninfo_by_id()
would return error on this session. If session has no connection
in this scenario, the session would still be residue because it
would not call in iscsi_sync_session().

No matter what is the root reason, it's an invalid session
which should be destroyed when it comes to iscsi_sync_session().

Signed-off-by: Wenchao Hao <haowenchao@huawei.com>