iscsid: destroy session and connection if connection's state is down
A connection would be in down state in next scenario:

- iscsid died after stop_conn() is called during logout
- iscsid died after create_conn() is called during login

For the first scenario, the connection is stopped with STOP_CONN_TERM,
it's a unrecoverable flag. For drivers which based on libiscsi, once
stop_conn() is triggerer, the connection can not send pdu any more.

While for the second scenario, iscsi_sysfs_get_sessioninfo_by_id()
would returned error, so it would not call in iscsi_sync_session(),
this session would still be residual.

Both the above scenario the session would failed to recovery, so
we need to destroy it.

Signed-off-by: Wenchao Hao <haowenchao@huawei.com>