More meson cleanup: handling older version, and version#
List our dependency on libpthred. Also, handle older version
of meson by checking the version number of meson before
using advanced features.

This means we don't have a default for "sbindir" now, but
that's okay, post-processing can move /usr/sbin binaries,
as needed.

Lastly, take the version number from meson.build and use
it to build the binaries and our library, so the version
number doesn't have to be in 3 places.