Small bug fixes (#364)
* build: move utils/sysdeps to top-level

The routines here are used by iscsiuio, so
move them to the top level.

* libopeniscsiusr: fix kernel-doc bugs

There were errors in four places, effecting the generated
manual pages.

* libopeniscsiusr: clean up kernel-doc usage

Simplify generation of library interface documentation using
the kernel-doc script:

* make the kernel-doc perl script executble (the only visible change)
* remove mention of sed (not used)
* stop compressing man pages -- let distros do that
* improve kernel-doc so that doc-preclean.pl and split-man.pl
  are no longer needed
* add list-man-pages.sh to simplify things (copied from libnvme)

* build: keep current version number in one place

Move the current version number (now 2.1.7) into
the top-level Makefile, and out of usr/version.h and
libopeniscsiusr/version.h, completely eliminating the
need for the latter.

Note that this ties the libopeniscsiusr library idea of
"version" to the usr/ (iscsid, etc) idea of version, even
though they used to be defined separately. But they always
had the same value, so nothing lost.

* build: insteall man pages uncompressed

We were delivering some man pages compressed and other not, and
it turns out distros know how to install man pages using the
compression they prefer, so stop doing it for them.